client/requirements.txt
getmac
requests
python-dotenv